+++ bumpr/vcs.py
@@ -1,5 +1,7 @@
 # -*- coding: utf-8 -*-
-from bumpr.helpers import execute
+from os.path import isdir
+
+from bumpr.helpers import execute, BumprError
 
 
 class BaseVCS(object):
@@ -7,38 +9,65 @@ class BaseVCS(object):
         self.verbose = verbose
 
     def execute(self, command):
+        '''Execute a command'''
         execute(command, verbose=self.verbose)
 
-    def commit(self, message, files):
+    def validate(self):
+        '''Ensure the working dir is a repository and there is no modified files'''
+        raise NotImplementedError
+
+    def commit(self, message):
+        '''Commit all modified files'''
         raise NotImplementedError
 
     def tag(self, name):
+        '''Create a tag'''
         raise NotImplementedError
 
 
 class Git(BaseVCS):
-    def commit(self, message, files):
-        for filename in files:
-            self.execute(["git", "add", filename])
-        self.execute(["git", "commit", "-m", message.encode('utf-8')])
+    def validate(self):
+        if not isdir('.git'):
+            raise BumprError('Current directory is not a git repopsitory')
+
+        for line in execute('git status --porcelain', verbose=False).splitlines():
+            if not line.startswith('??'):
+                raise BumprError('The current repository contains modified files')
+
+    def commit(self, message):
+        self.execute(["git", "commit", "-am", message.encode('utf-8')])
 
     def tag(self, name):
         self.execute(["git", "tag", name])
 
 
 class Mercurial(BaseVCS):
-    def commit(self, message, files):
-        for filename in files:
-            self.execute(["hg", "add", filename])
-        self.execute(["hg", "commit", "-m", message.encode('utf-8')])
+    def validate(self):
+        if not isdir('.hg'):
+            raise BumprError('Current directory is not a mercurial repopsitory')
+
+        for line in execute('hg status -mard', verbose=False).splitlines():
+            if not line.startswith('??'):
+                raise BumprError('The current repository contains modified files')
+
+    def commit(self, message):
+        self.execute(["hg", "commit", "-A", "-m", message.encode('utf-8')])
 
     def tag(self, name):
         self.execute(["hg", "tag", name])
 
 
 class Bazaar(BaseVCS):
-    def commit(self, message, files):
-        self.execute(["bzr"] + list(files) + ["commit", "-m", message.encode('utf-8')])
+    def validate(self):
+        if not isdir('.bzr'):
+            raise BumprError('Current directory is not a bazaar repopsitory')
+
+        for line in execute('bzr status --short', verbose=False).splitlines():
+            if not line.startswith('?'):
+                raise BumprError('The current repository contains modified files')
+
+    def commit(self, message):
+        self.execute(["bzr", "commit", "-m", message.encode('utf-8')])
 
     def tag(self, name):
         self.execute(["bzr", "tag", name])