+++ CHANGELOG.rst
@@ -4,7 +4,12 @@ Changelog
 Current
 -------
 
-- Nothing yet
+- Added ``--bump`` and ``--prepare`` to only perform bump or prepare
+- Rely on VCS for tracking files and ensure working copy is clean
+- Added option ``--nocommit``
+- Ensure dry run does not write or execute anything
+- Better output
+- Group parameters in help
 
 0.1.0 (2013-08-22)
 ------------------